refactor: Walkthrough
