+++ src/components/molecules/WalkThrough/Walkthrough.styled.tsx
