+++ src/components/molecules/WalkThrough/content/index.ts
@@ -1,0 +1,2 @@
+export {newReleaseFeaturesContent} from './newReleaseFeatures';
+export {noviceContent} from './novice';
