+++ src/components/molecules/WalkThrough/index.ts
@@ -1,1 +1,1 @@
-export {default} from './WalkThrough';
+export {default} from './Walkthrough';
