+++ src/components/molecules/index.ts
@@ -24,3 +24,4 @@ export {default as SectionRenderer} from './SectionRenderer';
 export {default as TemplateFormRenderer} from './TemplateFormRenderer';
 export {default as TitleBar} from './TitleBar';
 export {default as ValidationErrorsPopover} from './ValidationErrorsPopover';
+export {default as Walkthrough} from './Walkthrough';
