+++ src/models/walkthrough.ts
@@ -1,4 +1,5 @@
 export type WalkThroughCollection = 'novice' | 'release';
+
 export type WalkThroughStep<C extends WalkThroughCollection = WalkThroughCollection> = C extends 'novice'
   ? 'template' | 'resource' | 'syntax' | 'kustomizeHelm'
   : 'compare' | 'images';
